split each class formerly in plan_queue_ops.py into a specific file for each backend type. Additionally, split out the abstract class into its own py file.
